Refactor the context provider Fixes #271
